requirements.txt
pyserial
pyyaml
asyncio
aiofiles
numpy
paho-mqtt
getmac
pyyaml
escpos